corrected annee deletion

+++ projet-backend/Projet/Projet/src/main/java/com/fst/il/m2/Projet/controllers/AnneeController.java
@@ -1,8 +1,10 @@
 package com.fst.il.m2.Projet.controllers;
 
 import com.fst.il.m2.Projet.business.AnneeService;
+import com.fst.il.m2.Projet.business.DepartementService;
 import com.fst.il.m2.Projet.dto.AnneeDto;
 import com.fst.il.m2.Projet.mapper.AnneeMapper;
+import com.fst.il.m2.Projet.mapper.DepartementMapper;
 import com.fst.il.m2.Projet.models.Annee;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -16,10 +18,12 @@ import java.util.List;
 public class AnneeController {
 
     private final AnneeService anneeService;
+    private final DepartementService departementService;
 
     @Autowired
-    public AnneeController(AnneeService anneeService) {
+    public AnneeController(AnneeService anneeService, DepartementService departementService) {
         this.anneeService = anneeService;
+        this.departementService = departementService;
     }
 
     @PostMapping
@@ -46,7 +50,10 @@ public class AnneeController {
 
     @DeleteMapping("/{id}")
     public ResponseEntity<Void> deleteAnnee(@PathVariable Long id) {
-        anneeService.deleteAnnee(id);
-        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        if(!anneeService.hasDepartements(id)) {
+            anneeService.deleteAnnee(id);
+            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        }
+        return new ResponseEntity<>(HttpStatus.NOT_ACCEPTABLE);
     }
 }
